refactor: rename models folder, add info in readme
