+++ src/modules/album/album.schema.ts
