+++ src/modules/artist/artist.schema.ts
