+++ src/modules/band/band.resolver.ts
