+++ src/modules/band/band.schema.ts
