+++ src/modules/favorites/favorites.schema.ts
